LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/image-src.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""

    <!-- speculative case in document.write -->
    <image src="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=43ceaad4-6cc8-4561-afbd-fc621abccb40&amp;encodingcheck=&Gbreve;">


FAIL Speculative parsing, document.write(): image-src Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
